fix: cleanup

+++ solidity/interfaces/modules/IPrivateERC20ResolutionModule.sol
@@ -21,8 +21,8 @@ interface IPrivateERC20ResolutionModule is IResolutionModule {
 
   event VoteCommited(address _voter, bytes32 _disputeId, bytes32 _commitment);
   event VoteRevealed(address _voter, bytes32 _disputeId, uint256 _numberOfVotes);
-  event CommitingPhaseStarted(uint128 _startTime, bytes32 _disputeId);
-  event RevealingPhaseStarted(uint128 _startTime, bytes32 _disputeId);
+  event CommitingPhaseStarted(uint256 _startTime, bytes32 _disputeId);
+  event RevealingPhaseStarted(uint256 _startTime, bytes32 _disputeId);
   event DisputeResolved(bytes32 _disputeId, IOracle.DisputeStatus _status);
 
   error PrivateERC20ResolutionModule_OnlyDisputeModule();
